+++ crates/nostr/src/lib.rs
@@ -55,9 +55,7 @@ pub mod util;
 #[doc(hidden)]
 pub use self::event::tag::{Tag, TagKind, TagStandard, Tags};
 #[doc(hidden)]
-pub use self::event::{
-    Event, EventBuilder, EventId, Kind, MissingPartialEvent, PartialEvent, UnsignedEvent,
-};
+pub use self::event::{Event, EventBuilder, EventId, Kind, UnsignedEvent};
 #[doc(hidden)]
 pub use self::filter::{Alphabet, Filter, SingleLetterTag};
 #[doc(hidden)]